% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coords.R
\name{check_coords}
\alias{check_coords}
\title{Check site coordinates}
\usage{
check_coords(df)
}
\arguments{
\item{df}{Data frame with a column named in_out.}
}
\value{
a data frame.
}
\description{
Check site coordinates
}
