% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_in_out.R
\name{check_in_out}
\alias{check_in_out}
\title{Check values of in_out column}
\usage{
check_in_out(df)
}
\arguments{
\item{df}{Data frame with a column named in_out.}
}
\value{
a data frame.
}
\description{
Check that all values in in_out column are either 'inside' or 'outside'.
Otherwise, an error will report which cases need to be revised.
}
