% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sitenames.R
\name{check_sitenames}
\alias{check_sitenames}
\title{Check site names against sites list}
\usage{
check_sitenames(df)
}
\arguments{
\item{df}{Data frame with a column named site.}
}
\value{
a data frame.
}
\description{
If any site name does not match sitenames, an error
will report which these cases are so that they can be revised.
}
