% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spnames.R
\name{check_spnames}
\alias{check_spnames}
\title{Check species names against spp2exclude and spp2include}
\usage{
check_spnames(df, wide = FALSE)
}
\arguments{
\item{df}{Data frame with a column named species.}

\item{wide}{Logical. Data frame in wide format? Only TRUE for raw cover data (default is FALSE).
It wide = TRUE, species are assumed to start at the 5th column (until the end of the data frame).}
}
\value{
a data frame, possibly with some rows removed (those identified as species to exclude).
}
\description{
Check species names in a data frame. First, rows whose species name matches any in spp2exclude
are automatically removed from the data frame. Then remaining rows are checked
against spp2include: if any species name does not match spp2include, an error
will report which these cases are so that they can be revised.
}
