% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset.R
\name{make_dataset}
\alias{make_dataset}
\title{Prepare full dataset}
\usage{
make_dataset(spp = "data/species_info.csv",
  cover = "data/exclosure_cover_long.csv", damage = "data/damage.csv",
  sites = "data/sites_info.csv")
}
\arguments{
\item{spp}{Path to data frame.}

\item{cover}{Path to data frame.}

\item{damage}{Path to data frame.}

\item{sites}{Path to data frame.}
}
\value{
a data frame, and a csv file: \code{data/fulldata.csv}.
}
\description{
Prepare and assemble full dataset, after having processed all data components (using read_* functions).
}
