% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscatter.R
\name{multiscatter}
\alias{multiscatter}
\title{Produce multipanel scatterplot of cover/damage within and outside exclosures}
\usage{
multiscatter(dataset, vble, shrub.tree = "shrub", remove.zeros = TRUE,
  show.bare = TRUE, min.n = 1, xlab = "Outside", ylab = "Inside",
  fontsize = 12, psize = 4, n.row = NULL, n.col = NULL,
  filename = NULL, ...)
}
\arguments{
\item{dataset}{Name of the dataframe.}

\item{vble}{character. Either "cover" or "damage".}

\item{shrub.tree}{character. Either "shrub" (default) or "tree".}

\item{remove.zeros}{Logical. Show dots with cero cover both inside and outside? Default is TRUE.}

\item{show.bare}{Logical. Show bare ground cover as if another species (the default), or omit from the plot?}

\item{min.n}{Numeric. Omit species occurring at a number of sites lower than this. Default is 1, which implies plotting all species.}

\item{xlab}{character. Horizontal axis title.}

\item{ylab}{character. Vertical axis title.}

\item{psize}{numeric. Point size (for geom_point).}

\item{n.row, n.col}{Numeric. Number of rows/columns of the final plot. Default is NULL, which leaves ggplot choose the best display.}

\item{filename}{character (optional). If provided, will save the plot with this filename.}

\item{...}{further arguments for \code{\link[ggplot2]{ggsave}}, such as width, height, paper, etc.}
}
\value{
Prints (and optionally saves) a ggplot.
}
\description{
Produce multipanel scatterplot of cover/damage within and outside exclosures
}
