% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_damage.R
\name{read_damage}
\alias{read_damage}
\title{Read browsing damage data}
\usage{
read_damage(df)
}
\arguments{
\item{df}{Path to data frame}
}
\value{
a data frame.
}
\description{
Reads browsing damage data and check species names (see \code{\link{check_spnames}}).
}
