% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rawcover.R
\name{read_rawcover}
\alias{read_rawcover}
\title{Read raw cover data}
\usage{
read_rawcover(rawcover, tr.length)
}
\arguments{
\item{rawcover}{character. Path to a csv file containing the raw cover data. This data frame must fulfill certain conditions: (i) first four columns must be named \code{id, position, shrub_tree, section} and will be used to identify individual transects; (ii) Species raw cover data are stored in the 5th to nth column. See \code{\link[vegetools]{calculate_cover}} for more information about data formatting requirements.}

\item{tr.length}{Numeric. Transect length (if transect divided in sections, give section length).}
}
\value{
A data frame, and two csv files saved to the \code{data} folder:
\code{exclosure_cover_wide.csv} and \code{exclosure_cover_long.csv}.
}
\description{
Read and process raw cover data to calculate total cover per species per transect.
The function performs some minimal checks, e.g. will give an error if column names
do not match species names in \code{spp2include.csv}.
}
