% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_siteinfo.R
\name{read_siteinfo}
\alias{read_siteinfo}
\title{Read site information}
\usage{
read_siteinfo(df)
}
\arguments{
\item{df}{Path to data frame}
}
\value{
a data frame.
}
\description{
Reads site information and check site names (see \code{\link{check_sitenames}}).
}
