% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sppinfo.R
\name{read_sppinfo}
\alias{read_sppinfo}
\title{Read species info}
\usage{
read_sppinfo(sppdata, outdf = "data/species_info.csv")
}
\arguments{
\item{sppdata}{Path to csv file containing species data.}

\item{outdf}{Path to save cleaned csv.}
}
\value{
A dataframe, and a csv file saved to disk.
}
\description{
Read and prepare species information. Basically, create acronym for each species
using \code{\link[vegan]{make.cepnames}}. But note that 'Genista tridens' is
given acronym "Genitrids", and "Genista tridentata" is given acronym "Genitrid".
}
