% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_inout.R
\name{scatter_inout}
\alias{scatter_inout}
\title{Create scatterplot of number of species, cover or height inside vs outside exclosures}
\usage{
scatter_inout(dataset, vble, shrub.tree = "shrub", xlab = "Outside",
  ylab = "Inside", fontsize = 12, psize = 6, jitter = TRUE,
  filename = NULL, ...)
}
\arguments{
\item{dataset}{Name of the dataframe.}

\item{vble}{Character. Either "nspp", "cover", or "height".}

\item{shrub.tree}{character. Either "shrub" (default) or "tree".}

\item{xlab}{character. Horizontal axis title.}

\item{ylab}{character. Vertical axis title.}

\item{psize}{numeric. Point size (for geom_point).}

\item{jitter}{Logical.}

\item{filename}{character (optional). If provided, will save the plot with this filename.}

\item{...}{further arguments for \code{\link[ggplot2]{ggsave}}, such as width, height, paper, etc.}
}
\value{
Prints (and optionally saves) a ggplot.
}
\description{
Create scatterplot of number of species, cover or height inside vs outside exclosures
}
